Force inline some functions & shrink counters.
This allows much (~40%) better performance on Android, as various
functions that should be inlined weren't being inlined before
(presumably due gbcc being a shared library on Android, as opposed to an
executable on desktop).

+++ src/apu.h
@@ -77,8 +77,8 @@ struct channel {
 };
 
 struct apu {
-	uint64_t sync_clock;
-	uint64_t sample;
+	uint16_t sync_clock;
+	uint16_t sample;
 	uint8_t left_vol;
 	uint8_t right_vol;
 	bool disabled;
